Fix potential problems on server disconnect

+++ src/server/remote/remote-server.ts
@@ -74,6 +74,7 @@ export class RemoteServer extends Server<RemoteServerSettings> {
     }
 
     async onNetDisconnect() {
+        if (this.destroyed) return
         console.log('server disconnected')
         await multi.destroyAndStartLoop()
         sc.Dialogs.showErrorDialog('Disconnected')
